Hunk diff line selection: Now available in Safari
Allow Safari users to quote code in Safari, I guess
